Fix up integration tests
