+++ src/routes.rs
@@ -1,30 +1,42 @@
-use actix_web::*;
+use std::sync::Arc;
+
+use axum::extract::{Path, State};
+use axum::http::StatusCode;
+use axum::response::ErrorResponse;
+
+use axum::routing::{delete, get, patch, post};
+use axum::Router;
 use log::*;
 use serde::{Deserialize, Serialize};
-use sqlx::PgPool;
 use validator::Validate;
 
-use crate::{db, dto, route_error::BasicError};
+use crate::entity::{TodoTask, TodoUser};
+use crate::routing_utils::{DbErrorResponse, Json, ValidationErrorResponse};
+use crate::{db, dto, AppState, SharedData};
 
-// For a majority of the endpoints I don't use this macro because incomplete handlers generate imprecise error squiggles in the IDE.
 /// Sample endpoint that can be used to show the API is responsive.
-#[get("/")]
-pub async fn hello() -> impl Responder {
+pub async fn hello() -> &'static str {
     info!("Hello");
-    HttpResponse::Ok().body("Hello world!")
+    "Hello world!"
 }
 
-/// Attaches routes under "/users" to the application router.
-pub fn add_user_routes(config: &mut web::ServiceConfig) {
-    config
-        .route("/users", web::get().to(get_users))
-        .route("/users", web::post().to(create_user));
+/// Builds a router for all the user routes
+pub fn user_routes() -> Router<Arc<SharedData>> {
+    // config
+    //     .route("/users", web::get().to(get_users))
+    //     .route("/users", web::post().to(create_user));
+    Router::new()
+        .route("/", get(get_users))
+        .route("/", post(create_user))
+        .route("/:user_id/tasks", get(get_tasks_for_user))
+        .route("/:user_id/tasks/:task_id", get(get_task_for_user))
+        .route("/:user_id/tasks", post(add_task_for_user))
 }
 
 /// Retrieves a list of all the users in the system.
-async fn get_users(pg_pool: web::Data<PgPool>) -> Result<HttpResponse, BasicError> {
+async fn get_users(State(app_data): AppState) -> Result<Json<Vec<TodoUser>>, ErrorResponse> {
     info!("Requested users");
-    let db_cxn = pg_pool.get_ref();
+    let db_cxn = &app_data.db;
     let users_result = db::get_users(db_cxn).await;
     if users_result.is_err() {
         error!(
@@ -32,61 +44,58 @@ async fn get_users(pg_pool: web::Data<PgPool>) -> Result<HttpResponse, BasicErro
             users_result.as_ref().unwrap_err()
         );
     }
-    Ok(HttpResponse::Ok().json(users_result.map_err(BasicError::from_db)?))
+    let response = users_result.map_err(DbErrorResponse::from)?;
+    Ok(Json(response))
 }
 
 /// Creates a user.
 async fn create_user(
-    pg_pool: web::Data<PgPool>,
-    user_to_create: web::Json<dto::NewUser>,
-) -> Result<HttpResponse, BasicError> {
+    State(app_data): AppState,
+    Json(user_to_create): Json<dto::NewUser>,
+) -> Result<(StatusCode, Json<dto::InsertedUser>), ErrorResponse> {
     info!("Attempt to create user: {}", user_to_create);
     user_to_create
         .validate()
-        .map_err(BasicError::from_validate)?;
+        .map_err(ValidationErrorResponse::from)?;
 
-    let db_cxn = pg_pool.get_ref();
-    let creation_result = db::create_user(db_cxn, &user_to_create.into_inner()).await;
+    let db_cxn = &app_data.db;
+    let creation_result = db::create_user(db_cxn, &user_to_create).await;
     if creation_result.is_err() {
         error!(
             "User create failure: {}",
             creation_result.as_ref().unwrap_err()
         );
     }
-    Ok(HttpResponse::Created().json(
-        creation_result
-            .map_err(BasicError::from_db)
-            .map(|id| dto::InsertedUser { id })?,
+    Ok((
+        StatusCode::CREATED,
+        Json(
+            creation_result
+                .map_err(DbErrorResponse::from)
+                .map(|id| dto::InsertedUser { id })?,
+        ),
     ))
 }
 
 /// Adds routes under "/tasks" and routes for user-owned tasks to the application router
-pub fn add_task_routes(config: &mut web::ServiceConfig) {
-    config
-        .route("/users/{user_id}/tasks", web::get().to(get_tasks_for_user))
-        .route(
-            "/users/{user_id}/tasks/{task_id}",
-            web::get().to(get_task_for_user),
-        )
-        .route("/users/{user_id}/tasks", web::post().to(add_task_for_user))
-        .route("/tasks/{task_id}", web::patch().to(update_task_for_user))
-        .route("/tasks/{task_id}", web::delete().to(delete_task_for_user));
+pub fn task_routes() -> Router<Arc<SharedData>> {
+    Router::new()
+        .route("/tasks/:task_id", patch(update_task))
+        .route("/tasks/:task_id", delete(delete_task))
 }
 
 /// Retrieves a set of tasks owned by a user
 async fn get_tasks_for_user(
-    pg_pool: web::Data<PgPool>,
-    user_id: web::Path<i32>,
-) -> Result<HttpResponse, BasicError> {
+    State(app_state): AppState,
+    Path(user_id): Path<i32>,
+) -> Result<Json<Vec<TodoTask>>, ErrorResponse> {
     info!("Get tasks for user {user_id}");
-    let db_cxn = pg_pool.get_ref();
-    let unwrapped_user_id = user_id.into_inner();
-    let tasks = db::get_tasks_for_user(db_cxn, unwrapped_user_id).await;
+    let db_cxn = &app_state.db;
+    let tasks = db::get_tasks_for_user(db_cxn, user_id).await;
     if tasks.is_err() {
         let err = tasks.as_ref().unwrap_err();
-        error!("Failed to get user {unwrapped_user_id}'s tasks: {}", err)
+        error!("Failed to get user {user_id}'s tasks: {}", err)
     }
-    Ok(HttpResponse::Ok().json(tasks.map_err(BasicError::from_db)?))
+    Ok(Json(tasks.map_err(DbErrorResponse::from)?))
 }
 
 #[derive(Deserialize)]
@@ -97,11 +106,11 @@ struct GetTaskPath {
 
 /// Retrieves a specific task owned by a user
 async fn get_task_for_user(
-    pg_pool: web::Data<PgPool>,
-    path: web::Path<GetTaskPath>,
-) -> Result<HttpResponse, BasicError> {
+    State(app_state): AppState,
+    Path(path): Path<GetTaskPath>,
+) -> Result<Json<TodoTask>, ErrorResponse> {
     info!("Get task {} for user {}", path.task_id, path.user_id);
-    let db_cxn = pg_pool.get_ref();
+    let db_cxn = &app_state.db;
     let task = db::get_task_for_user(db_cxn, path.user_id, path.task_id).await;
     if let Err(ref contained_err) = task {
         // We don't want to log an error for the "no results" case
@@ -113,7 +122,8 @@ async fn get_task_for_user(
             ),
         }
     }
-    Ok(HttpResponse::Ok().json(task.map_err(BasicError::from_db)?))
+
+    Ok(Json(task.map_err(DbErrorResponse::from)?))
 }
 
 #[derive(Serialize)]
@@ -123,64 +133,69 @@ struct InsertedTask {
 
 /// Adds a new task for a user
 async fn add_task_for_user(
-    pg_pool: web::Data<PgPool>,
-    user_id: web::Path<i32>,
-    task_data: web::Json<dto::NewTask>,
-) -> Result<HttpResponse, BasicError> {
+    State(app_state): AppState,
+    Path(user_id): Path<i32>,
+    Json(task_data): Json<dto::NewTask>,
+) -> Result<(StatusCode, Json<InsertedTask>), ErrorResponse> {
     info!("Adding task for user {user_id}");
-    task_data.validate().map_err(BasicError::from_validate)?;
+    task_data
+        .validate()
+        .map_err(ValidationErrorResponse::from)?;
 
-    let db_cxn = pg_pool.get_ref();
-    let unwrapped_user_id = user_id.into_inner();
-    let inserted_task =
-        db::add_task_for_user(db_cxn, unwrapped_user_id, &task_data.into_inner()).await;
+    let db_cxn = &app_state.db;
+    let inserted_task = db::add_task_for_user(db_cxn, user_id, &task_data).await;
     if inserted_task.is_err() {
         error!(
-            "Failed to add task for user {unwrapped_user_id}: {}",
+            "Failed to add task for user {user_id}: {}",
             inserted_task.as_ref().unwrap_err()
         );
     }
-    Ok(HttpResponse::Created().json(
-        inserted_task
-            .map_err(BasicError::from_db)
-            .map(|id| InsertedTask { id })?,
+
+    Ok((
+        StatusCode::CREATED,
+        Json(
+            inserted_task
+                .map_err(DbErrorResponse::from)
+                .map(|id| InsertedTask { id })?,
+        ),
     ))
 }
 
 /// Updates the content of a task
-async fn update_task_for_user(
-    pg_pool: web::Data<PgPool>,
-    task_id: web::Path<i32>,
-    task_data: web::Json<dto::UpdateTask>,
-) -> Result<HttpResponse, BasicError> {
+async fn update_task(
+    State(app_state): AppState,
+    Path(task_id): Path<i32>,
+    Json(task_data): Json<dto::UpdateTask>,
+) -> Result<StatusCode, ErrorResponse> {
     info!("Updating task {task_id}");
-    task_data.validate().map_err(BasicError::from_validate)?;
+    task_data
+        .validate()
+        .map_err(ValidationErrorResponse::from)?;
 
-    let db_cxn = pg_pool.get_ref();
-    let update_result =
-        db::update_user_task(db_cxn, task_id.into_inner(), &task_data.into_inner()).await;
+    let db_cxn = &app_state.db;
+    let update_result = db::update_user_task(db_cxn, task_id, &task_data).await;
     match update_result {
-        Ok(_) => Ok(HttpResponse::Ok().finish()),
+        Ok(_) => Ok(StatusCode::OK),
         Err(db_err) => {
             error!("Update task failure: {db_err}");
-            Err(BasicError::from_db(db_err))
+            Err(DbErrorResponse::from(db_err).into())
         }
     }
 }
 
 /// Deletes a task
-async fn delete_task_for_user(
-    pg_pool: web::Data<PgPool>,
-    task_id: web::Path<i32>,
-) -> Result<HttpResponse, BasicError> {
+async fn delete_task(
+    State(app_state): AppState,
+    Path(task_id): Path<i32>,
+) -> Result<StatusCode, ErrorResponse> {
     info!("Deleting task {task_id}");
-    let db_cxn = pg_pool.get_ref();
-    let delete_result = db::delete_user_task(db_cxn, task_id.into_inner()).await;
+    let db_cxn = &app_state.db;
+    let delete_result = db::delete_user_task(db_cxn, task_id).await;
     match delete_result {
-        Ok(_) => Ok(HttpResponse::Ok().finish()),
+        Ok(_) => Ok(StatusCode::OK),
         Err(db_err) => {
             error!("Failed to delete task: {db_err}");
-            Err(BasicError::from_db(db_err))
+            Err(DbErrorResponse::from(db_err).into())
         }
     }
 }
